requirements.txt
google-play-scraper
pandas
google-generativeai